sysctl: Get rid of GOOS targets
It doesn't make sense to have the dummy Darwin GOOS targets to make
Cilium compile on OS X.

Instead, get rid of GOOS targets, and just have a single for Linux
without any build constraints. The sysctl linux module doesn't have
any Linux-specific calls which would not compile on OS X.
