+++ pkg/sysctl/sysctl_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build linux
+// +build !privileged_tests
 
 package sysctl
